SIEM_Project/log_collector.py
import os
import time
import json
from datetime import datetime
import logging
from pathlib import Path
import Evtx.Evtx as evtx

class LogCollector:
    def __init__(self, log_directory="logs"):
        self.log_directory = log_directory
        self.setup_logging()
        
    def setup_logging(self):
        logging.basicConfig(
            level=logging.INFO,
            format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
            handlers=[
                logging.FileHandler('siem.log'),
                logging.StreamHandler()
            ]
        )
        self.logger = logging.getLogger('SIEM_Collector')

    def collect_windows_events(self, evtx_path):
        """Collect Windows Event logs"""
        try:
            with evtx.Evtx(evtx_path) as log:
                for record in log.records():
                    yield {
                        'timestamp': record.timestamp().isoformat(),
                        'event_id': record.event_id(),
                        'data': record.xml()
                    }
        except Exception as e:
            self.logger.error(f"Error processing Windows Event log {evtx_path}: {str(e)}")

    def collect_system_logs(self):
        """Collect system logs"""
        # Add your system log collection logic here
        pass

    def collect_application_logs(self):
        """Collect application logs"""
        # Add your application log collection logic here
        pass

    def process_logs(self):
        """Process collected logs"""
        try:
            # Windows Event Logs
            windows_logs = "C:\\Windows\\System32\\winevt\\Logs"
            security_log = os.path.join(windows_logs, "Security.evtx")
            
            if os.path.exists(security_log):
                self.logger.info("Processing Windows Security logs...")
                for event in self.collect_windows_events(security_log):
                    # Process and store events
                    self.logger.info(f"Collected event: {event['event_id']}")
            
            # Add more log sources here
            
        except Exception as e:
            self.logger.error(f"Error in log processing: {str(e)}")

if __name__ == "__main__":
    collector = LogCollector()
    collector.process_logs()
