SIEM_Project/requirements.txt
python -m venv venv
# Activate the virtual environment
# On Windows:
venv\Scripts\activate
# Then install requirements
pip install -r requirements.txtpython-dateutil==2.8.2
elasticsearch==8.9.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
python-evtx==0.7.4
pytz==2023.3
requests==2.31.0
python-dotenv==1.0.0
colorama==0.4.6
